Fixed many usage issues.  And CSS.

+++ app/controllers/user_sessions_controller.rb
@@ -1,6 +1,5 @@
 class UserSessionsController < ApplicationController
   before_filter :not_logged_in_required, :only => [ :new, :create ]
-  before_filter :login_required, :only => :destroy
 
   def new
     @user_session = UserSession.new
@@ -12,15 +11,17 @@ class UserSessionsController < ApplicationController
     @user_session = UserSession.new(params[:user_session])
     if @user_session.save
       flash[:notice] = t(:logged_in, :scope => [ :authentication ])
-      redirect_back_or_default home_path
+      redirect_back_or_default Configuration.default_path
     else
       render :action => :new
     end
   end
 
   def destroy
-    current_user_session.destroy
-    flash[:notice] = t(:logged_out, :scope => [ :authentication ])
+    if logged_in?
+      current_user_session.destroy
+      flash[:notice] = t(:logged_out, :scope => [ :authentication ])
+    end
     redirect_back_or_default login_path
   end
 end
